+++ src/app/_components/slider/index.ts
