+++ src/app/_features/product/detail/components/ProductDetail.tsx
