+++ src/app/_features/product/detail/endpoint.ts
