+++ src/app/_features/product/form/endpoint.ts
@@ -1,4 +1,4 @@
-import { supabase } from "app/_lib/supabase";
+import { supabase } from "src/app/_lib/supabase";
 import { Product } from "../detail/endpoint";
 import axios from "axios";
 
@@ -7,14 +7,14 @@ export const postImage = async (file: File, fileName: string) => {
     .from("product-image-buckets")
     .upload(`thumbnail/${fileName}`, file);
   return error || null;
-}
+};
 
 export const getImageUrl = async (fileName: string) => {
   const { data: urlData } = await supabase.storage
     .from("product-image-buckets")
     .getPublicUrl(`thumbnail/${fileName}`);
   return urlData.publicUrl;
-}
+};
 
 export type PostProductRequest = Omit<Product, "productCUID">;
 
@@ -23,6 +23,6 @@ export const postProduct = async (body: PostProductRequest) => {
     method: "POST",
     url: "/api/product",
     data: body,
-  })
+  });
   return response.status === 201;
-}
+};
