+++ src/app/_features/product/list/endpoint.ts
