+++ src/app/_features/product/list/hooks.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { getProductList } from "./endpoint";
-import { productsKeys } from "app/_utils";
+import { productsKeys } from "src/app/_utils";
 
 // 認証なしだとGET APIは動作しないため、actionsを使ってGETしている
 // そのため、この関数は使っていない
