+++ src/app/product/[productCUID]/page.tsx
