+++ tsconfig.json
@@ -19,10 +19,10 @@
       }
     ],
     "paths": {
-      "@components/*": ["app/_components/*"],
-      "@hooks": ["app/_utils"],
-      "@mock": ["app/_mock"],
-      "@types": ["app/_types"]
+      "@components/*": ["src/app/_components/*"],
+      "@hooks": ["src/app/_utils"],
+      "@mock": ["src/app/_mock"],
+      "@types": ["src/app/_types"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
